requirements.txt
# LLM + Orchestration
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
openai>=1.40.0          # OpenAI-compatible client for Qwen-max via DashScope

# Vector DB
chromadb>=0.5.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0

# Frontend
gradio>=5.0.0
pyvis>=0.3.2

# Config + Utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tiktoken>=0.7.0
networkx>=3.3

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
